Add a db manager and enable the use of pipelines
Use the singleton  pattern to create an one-access point to the db.

Thus, modify the spider and the pipelines to use the db manager.

Get rid of the filmTitle attribute both in the spider and the item container, since the films are added from the spider just after reading their infos.