Set to store the reviews in a mysql db
Modify the spider and the db module to better match the use of mysql.

Add a python mysql connector to the to be installed modules. in the README.
Update it with also a tutorial to set up the db workspace.

Set an upper limit for the String type columns in the db module, since mysql requires the length for CHAR and VARCHAR types.
Modify also the URL to connect to via the pymysql driver and through the access data set before.

If a film is already in the db, adding it again raises an error and not let the reviews to be added. Get rid of it by handling with a try-catch statement in the spider.