Modify the structure by adding a co-fetcher class
The new class helps the spider to correctly fetch reviews for each film.

fimlID and filmTitle attributes changed quicker than the request were yielded, so it resulted to link some reviews to one film, instead of the correct one.
Get rid of this problem by shifting the parsing method to the spider's helper with a copy reference of the film's ID and title, and the url request, so that each instance of the helper refers to a single film in the dataset.